Use NSInteger tag.

+++ Palette/NSView+Spinny.m
@@ -17,9 +17,9 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 @interface HackedProgressIndicator : NSProgressIndicator
 {
-  int _tag;  
+  NSInteger _tag;
 }
-@property (assign) int tag;
+@property (assign) NSInteger tag;
 @end
 
 @implementation HackedProgressIndicator
